refactored codegen to make it more clear what was going on
