+++ src/components/HeadingIndicator.tsx
