+++ src/components/Instrument.tsx
